GUI.cpp and GUI.h created IMGUI added to GUI
