+++ Application/imgui.ini
@@ -4,17 +4,16 @@ Size=400,400
 Collapsed=0
 
 [Window][Dear ImGui Demo]
-Pos=171,0
-Size=550,680
+Pos=427,7
+Size=550,648
 Collapsed=0
-DockId=0x00000001,0
+DockId=0x00000002,0
 
 [Window][Hello, world!]
-Pos=210,20
-Size=550,680
+Pos=863,160
+Size=276,648
 Collapsed=0
-DockId=0x00000001,1
 
 [Docking][Data]
-DockNode  ID=0x00000001 Pos=210,20 Size=550,680 Selected=0xD00C9CD4
+DockNode  ID=0x00000002 Pos=427,7 Size=550,648 Selected=0xE87781F4
