+++ Application/src/Application.cpp
@@ -1,0 +1,6 @@
+#include "GUI.h"
+
+int main()
+{
+
+}
